Move "OfficeTextField"'s types to new file Add prefix and suffix to "OfficeTextField"
